// src/app/models/employee.ts
export interface employee {
    emp_no: number;
    full_name: string;
    part_full_time_flag: string; 
    religion: string;
    emp_sex: string;
    blood_group: string;
    marital_status: string;
    marital_dated: Date;
    date_of_birth: Date;
    country_of_birth: string;
    city_of_birth: string;
    can_travel_local: string;
    can_travel_abroad: string;
    mobile_no: string;
    official_email_add: string;
    personal_email_add: string;
    totalRecords: number;
}